fix(zc)!: move the title screen logo smoothly instead of in whole-pixel steps
end changelog

The logo crawl (and the sword thrust) animate at a fractional number of
game pixels per frame, but render positions were stored and drawn as
integers - so the motion advanced in whole game-pixel jumps of uneven
cadence, several output pixels at a time once scaled up to the window.

Render tree transforms now store x/y as floats, and item draws map their
corners through a float variant of the matrix apply, so sub-pixel
positions survive to the draw call (and to the CRT overlay pass's
sub-rect). Mouse mapping and layout code keep their integer paths;
existing call sites pass the same whole numbers as before, just cast.

+++ src/launcher/launcher.cpp
@@ -303,16 +303,16 @@ static void configure_render_tree()
 			yscale = std::max((int) yscale, 1);
 		}
 		rti_screen.set_transform({
-			.x = (int)(resx - w*xscale) / 2,
-			.y = (int)(resy - h*yscale) / 2,
+			.x = (float)((int)(resx - w*xscale) / 2),
+			.y = (float)((int)(resy - h*yscale) / 2),
 			.xscale = xscale,
 			.yscale = yscale,
 		});
 		rti_screen.visible = true;
 		// TODO: it seems `rti_screen` is unnecessary, given this `rti_dialogs` also draws the main Window component.
 		rti_dialogs.set_transform({
-			.x = (int)(resx - w*xscale) / 2,
-			.y = (int)(resy - h*yscale) / 2,
+			.x = (float)((int)(resx - w*xscale) / 2),
+			.y = (float)((int)(resy - h*yscale) / 2),
 			.xscale = xscale,
 			.yscale = yscale,
 		});
